--- Geometry/TrackerNumberingBuilder/plugins/TrackerGeometricDetExtraESModule.h
@@ -1,4 +1,0 @@
-#ifndef Geometry_TrackerNumberingBuilder_TrackerGeometricDetExtraESModule_H
-#define Geometry_TrackerNumberingBuilder_TrackerGeometricDetExtraESModule_H
-
-#endif
